Fixed one runtime and hopefully one compile-time bug for GCC in the parallel mode.
